Implement working APIs
- Set up an express server
- Get handlers on 2 url routes
- '/images/:id' will return the image with corresponding id
- '/restaurant/:restaurant_id' will return all images with matching restaurant_id
