+++ database/index.js
@@ -1,10 +1,10 @@
 const Sequelize = require('sequelize');
-const generateData = require('./generateData.js');
 
 // create the sequelize instance
 const sequelize = new Sequelize('photo_gallery', 'root', '', {
   host: 'localhost',
-  dialect: 'mysql'
+  dialect: 'mysql',
+  // logging: false
 });
 
 sequelize
@@ -60,34 +60,34 @@ const Restaurants = sequelize.define('restaurants', {
   timestamps: false
 });
 
-// Generate random users, restaurants and images
-const randomUsers = generateData.generateUsers();
-const randomRestaurants = generateData.generateRestaurants();
-const randomImages = generateData.generateImages();
-
 Users.sync();
 Restaurants.sync();
 Images.sync();
 
+const retrieveImage = function (imageId, callback) {
+  Images.findOne({
+    where: {
+      id: imageId
+    }
+  }).then((result) => {
+    callback(result);
+  });
+};
 
-// Insert random data into corresponding database
-for (let i = 0; i < randomUsers.length; i++) {
-  Users.create(randomUsers[i])
-    .then((data) => {
-      console.log('User added to database', randomUsers[i].name);
-    });
-}
+const retrieveRestaurantImage = function (restaurantId, callback) {
+  Images.findAll({
+    where: {
+      restaurant_id: restaurantId
+    }
+  }).then((result) => {
+    callback(result);
+  });
+};
 
-for (let i = 0; i < randomRestaurants.length; i++) {
-  Restaurants.create(randomRestaurants[i])
-    .then((data) => {
-      console.log('Restaurant added to database', randomRestaurants[i].name);
-    });
-}
-
-for (let i = 0; i < randomImages.length; i++) {
-  Images.create(randomImages[i])
-    .then((data) => {
-      console.log('Image added to database', randomImages[i].name);
-    });
-}
+module.exports = {
+  Users,
+  Restaurants,
+  Images,
+  retrieveImage,
+  retrieveRestaurantImage
+};
