+++ package.json
@@ -9,10 +9,13 @@
   },
   "scripts": {
     "react-dev": "webpack -d --watch",
+    "server-dev": "node server/index.js",
     "test": "jest",
-    "generate-data": "node database/index.js"
+    "generate-data": "node database/generateData.js"
   },
   "dependencies": {
+    "body-parser": "^1.19.0",
+    "express": "^4.17.0",
     "faker": "^4.1.0",
     "jest": "^24.8.0",
     "mysql": "^2.17.1",
